Add files via upload
This is an HTML file with Javascript making use of WebGL and some of the functions that we can use with it. 
e.g adding a cube that is rotating on the render, controls, to make the user set at what speed the cube must rotate and also its size.